feat: ダウンロードボタンのcomponent分割
